+++ cmd/run.go
@@ -2,11 +2,15 @@ package cmd
 
 import (
 	"fmt"
+	"os"
+	"os/signal"
+	"syscall"
 	"time"
 
 	"github.com/lunemec/eve-quartermaster/pkg/bot"
 	"github.com/lunemec/eve-quartermaster/pkg/repository"
 	"github.com/lunemec/eve-quartermaster/pkg/token"
+	"github.com/pkg/errors"
 
 	"github.com/bwmarrin/discordgo"
 	"github.com/spf13/cobra"
@@ -45,7 +49,7 @@ func init() {
 	runCmd.Flags().Int32Var(&allianceID, "alliance_id", 0, "Alliance ID for which to list contracts")
 	runCmd.Flags().DurationVar(&checkInterval, "check_interval", 30*time.Minute, "how often to check EVE ESI API (default 30min)")
 	runCmd.Flags().DurationVar(&notifyInterval, "notify_interval", 24*time.Hour, "how often to spam Discord (default 24H)")
-	runCmd.Flags().StringVar(&repositoryFile, "repository_file", "repository.json", "path to repository json to save require_stock data (default repository.json)")
+	runCmd.Flags().StringVar(&repositoryFile, "repository_file", "repository.db", "path to bbolt repository file to save doctrine data (default repository.db)")
 
 	must(runCmd.MarkFlagRequired("session_key"))
 	must(runCmd.MarkFlagRequired("eve_client_id"))
@@ -55,6 +59,10 @@ func init() {
 }
 
 func runBot(cmd *cobra.Command, args []string) {
+	errChan := make(chan error)
+	signalChan := make(chan os.Signal, 1)
+	signal.Notify(signalChan, syscall.SIGINT, syscall.SIGTERM)
+
 	fastLog, err := zap.NewDevelopment()
 	if err != nil {
 		panic(fmt.Sprintf("error inicializing logger: %s", err))
@@ -80,10 +88,16 @@ func runBot(cmd *cobra.Command, args []string) {
 		panic(fmt.Sprintf("error inicializing discord client: %+v", err))
 	}
 
-	repository, err := repository.NewJSONRepository(repositoryFile)
+	repository, err := repository.NewBBoltRepository(repositoryFile)
 	if err != nil {
 		panic(fmt.Sprintf("error inicializing repository file: %+v", err))
 	}
+	defer func() {
+		err := repository.Close()
+		if err != nil {
+			fmt.Printf("ERROR closing DB: %+v\n", err)
+		}
+	}()
 
 	bot := bot.NewQuartermasterBot(
 		log,
@@ -97,9 +111,28 @@ func runBot(cmd *cobra.Command, args []string) {
 		checkInterval,
 		notifyInterval,
 	)
-	err = bot.Bot()
-	// systemd handles reload, so we can panic on error.
-	if err != nil {
-		panic(err)
+
+	go func() {
+		err := bot.Bot()
+		errChan <- err
+	}()
+
+	select {
+	case <-signalChan:
+		// Save bbolt DB to disk.
+		err := repository.Close()
+		if err != nil {
+			panic(errors.Wrap(err, "ERROR closing DB"))
+		}
+		// This forces us to refresh token + save to file.
+		_, err = tokenSource.Token()
+		if err != nil {
+			panic(errors.Wrap(err, "error refreshing and saving token"))
+		}
+	case <-errChan:
+		// systemd handles reload, so we can panic on error.
+		if err != nil {
+			panic(err)
+		}
 	}
 }