+++ pkg/handler/handler.go
@@ -47,7 +47,7 @@ func init() {
 
 // New constructs new API http handler.
 func New(signalChan chan os.Signal, log handlerLogger, client *http.Client, tokenStorage token.Storage, secretKey []byte, clientID, ssoSecret string, callbackURL string, scopes []string) http.Handler {
-	esi := goesi.NewAPIClient(client, "EVE Scanner (lukas@nemec.lu)")
+	esi := goesi.NewAPIClient(client, "EVE Quartermaster (lu.nemec@gmail.com)")
 	sso := goesi.NewSSOAuthenticatorV2(client, clientID, ssoSecret, callbackURL, scopes)
 	r := chi.NewRouter()
 	h := handler{
@@ -94,7 +94,7 @@ func (h *handler) tokenSource(r *http.Request, w http.ResponseWriter) (oauth2.To
 	ts := h.sso.TokenSource(&token)
 	newToken, err := ts.Token()
 	if err != nil {
-		return nil, errors.Errorf("error getting token")
+		return nil, errors.Wrapf(err, "error getting token")
 	}
 
 	if token != *newToken {